adding to add to MyFOodList works

+++ src/main/java/com/henrik/calorieapphenrik/food/dto/FoodDto.java
@@ -15,7 +15,7 @@ public class FoodDto {
 
 
     private String name;
-    private BigDecimal calories;
+    private Integer calories;
     private Integer protein;
     private Integer carbs;
     private Integer fat;
